Migrate posts to a different component
